back-end/requirements.txt
fastapi[standard]
uvicorn==0.34.0
aiogram==3.19.0
hydrogram==0.2.0
python-dotenv==1.1.0
asyncpg==0.30.0
tortoise-orm==0.24.2
TgCrypto==1.2.5
miniopy-async==1.22.1
urllib3
